// src/types/api.ts
/**
 * This file was auto-generated by openapi-typescript.
 * Do not make direct changes to the file.
 */

export interface paths {
  "/api/v1/health": {
    get: {
      responses: {
        200: {
          content: {
            "application/json": components["schemas"]["internal_api.HealthResponse"];
          };
        };
      };
    };
  };
  "/api/v1/agent/metrics": {
    get: {
      responses: {
        200: {
          content: {
            "application/json": components["schemas"]["internal_api.MetricsResponse"];
          };
        };
      };
    };
  };
  "/api/v1/agent/start": {
    post: {
      responses: {
        200: {
          content: {
            "application/json": components["schemas"]["internal_api.AgentActionResponse"];
          };
        };
      };
    };
  };
  "/api/v1/agent/stop": {
    post: {
      responses: {
        200: {
          content: {
            "application/json": components["schemas"]["internal_api.AgentActionResponse"];
          };
        };
      };
    };
  };
}

export interface webhooks {}

export interface components {
  schemas: {
    "internal_api.HealthResponse": {
      status?: string;
      uptime?: string;
      version?: string;
      links?: components["schemas"]["internal_api.Link"][];
    };
    "internal_api.AgentActionResponse": {
      message?: string;
      status?: string;
    };
    "internal_api.MetricsResponse": {
      data?: components["schemas"]["monitoring-system_internal_metrics.Metrics"];
      timestamp?: string;
    };
    "internal_api.Link": {
      href?: string;
      method?: string;
      rel?: string;
    };
    "monitoring-system_internal_metrics.Metrics": {
      hardware?: Record<string, unknown>;
      software?: Record<string, unknown>;
      system?: Record<string, unknown>;
    };
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}

export type $defs = never;